requirements.txt
fastapi
uvicorn
pydantic
selenium
webdriver-manager
python-dotenv
python-multipart
